+++ src/content.ts
@@ -2,7 +2,6 @@ import { requireGetCurrentFromURI, requireGetStore, ConfigOpts } from "./core/co
 import {
   requireValidActivityContentNote,
   requireValidActivityContentProfile,
-  serialize,
   ActivityContentNote,
   ActivityContentProfile,
 } from "./core/activityContent";
@@ -46,7 +45,7 @@ export const broadcast = async (
   opts?: ConfigOpts
 ): Promise<SignedBroadcastAnnouncement> => {
   requireValidActivityContentNote(contentObject);
-  const content = serialize(contentObject);
+  const content = JSON.stringify(contentObject);
 
   const currentFromURI = requireGetCurrentFromURI(opts);
 
@@ -91,7 +90,7 @@ export const reply = async (
   if (!isDSNPAnnouncementURI(inReplyTo)) throw new InvalidAnnouncementUriError(inReplyTo);
 
   requireValidActivityContentNote(contentObject);
-  const content = serialize(contentObject);
+  const content = JSON.stringify(contentObject);
 
   const currentFromURI = requireGetCurrentFromURI(opts);
 
@@ -157,7 +156,7 @@ export const profile = async (
   opts?: ConfigOpts
 ): Promise<SignedProfileAnnouncement> => {
   requireValidActivityContentProfile(contentObject);
-  const content = serialize(contentObject);
+  const content = JSON.stringify(contentObject);
 
   const currentFromURI = requireGetCurrentFromURI(opts);